+++ src/ocean/util/container/map/model/BucketInfo.d
@@ -19,8 +19,6 @@ module ocean.util.container.map.model.BucketInfo;
 
 *******************************************************************************/
 
-import ocean.core.Array: clear;
-
 debug (BucketInfo) import tango.io.Stdout;
 
 /*******************************************************************************/
@@ -37,11 +35,13 @@ class BucketInfo
     {
         /**********************************************************************
 
-            Bucket index, the index of the bucket in the array of buckets.
+            Bucket index, the index of the bucket in the array of buckets. It
+            is meaningful only if length > 0 so for easier bug detection the
+            initial value is an out-of-range index.
 
          **********************************************************************/
 
-        size_t index;
+        size_t index = size_t.max;
 
         /**********************************************************************
 
@@ -50,7 +50,7 @@ class BucketInfo
 
          **********************************************************************/
 
-        size_t length;
+        size_t length = 0;
 
         /**********************************************************************
 
@@ -85,6 +85,15 @@ class BucketInfo
         List of Bucket info instances, with the non-empty buckets first, so that
         buckets[0 .. n_filled] refers to the non-empty buckets.
 
+        All elements in buckets[n_filled .. $] must always have the initial
+        value Bucket.init so that, given a bucket index b, if
+
+            bucket_list_indices[b] >= n_filled
+
+        then
+
+            buckets[bucket_list_indices[b]].length == 0.
+
      **************************************************************************/
 
     private Bucket[] buckets;
@@ -94,6 +103,12 @@ class BucketInfo
         Index in the list of Bucket info instances by bucket index, so that for
         a bucket index b the bucket info for this bucket can be obtained by
         buckets[bucket_list_indices[b]].
+
+        All elements of this array are initialised to buckets.length - 1
+        because, as described for the buckets array, given a bucket index b that
+        refers to an empty bucket, bucket_list_indices[b] must be at least
+        n_filled, and as long as there are empty buckets, n_filled is less than
+        buckets.length.
 
      **************************************************************************/
 
@@ -123,6 +138,10 @@ class BucketInfo
         {
             assert (this.n_filled);
         }
+        else
+        {
+            assert(!this.n_filled);
+        }
     }
 
     /**************************************************************************
@@ -138,6 +157,12 @@ class BucketInfo
     {
         this.buckets             = new Bucket[num_buckets];
         this.bucket_list_indices = new size_t[num_buckets];
+
+        /*
+         * Initialise bucket_list_indices, see bucket_list_indices documentation
+         * for details.
+         */
+        this.bucket_list_indices[] = this.buckets.length - 1;
     }
 
 
@@ -291,10 +316,14 @@ class BucketInfo
     {
         with (this.buckets[this.n_filled])
         {
+            assert(index >= this.buckets.length);
             index  = bucket_index;
+
+            assert(!length);
             length = 1;
         }
 
+        assert(this.bucket_list_indices[bucket_index] >= this.n_filled);
         this.bucket_list_indices[bucket_index] = this.n_filled++;
 
         this.n_elements++;
@@ -319,11 +348,11 @@ class BucketInfo
     {
         assert (this);
 
+        assert (this.n_elements, "update: no element in map");
+
         assert (this.buckets[this.bucket_list_indices[bucket_index]].length,
                 "attempted to update an empty bucket info: use create()/put() "
                 "instead");
-
-        assert (this.n_elements, "update: no element in map");
 
         debug (BucketInfo) this.print("upd ", bucket_index);
     }
@@ -361,10 +390,10 @@ class BucketInfo
     {
         assert (this);
 
+        assert (this.n_elements, "remove: no element in map");
+
         assert (this.buckets[this.bucket_list_indices[bucket_index]].length,
                 "remove: attempted to remove an element from an empty bucket");
-
-        assert (this.n_elements, "remove: no element in map");
 
         debug (BucketInfo) this.print("rem ", bucket_index);
     }
@@ -413,11 +442,7 @@ class BucketInfo
                 this.bucket_list_indices[info_to_remove.index] = *bucket_info_index;
             }
 
-            with (*last_info)
-            {
-                index = this.buckets.length;
-                length = 0;
-            }
+            *last_info = (*last_info).init;
 
             *bucket_info_index = this.buckets.length;
         }
@@ -485,7 +510,15 @@ class BucketInfo
     }
     body
     {
-        .clear(this.filled_buckets);
+        /*
+         * Reset all buckets that have been in use.
+         */
+        this.filled_buckets[] = Bucket.init;
+        /*
+         * Reinitialise bucket_list_indices, see bucket_list_indices
+         * documentation for details.
+         */
+        this.bucket_list_indices[] = this.buckets.length - 1;
 
         this.n_filled = this.n_elements = 0;
     }
@@ -502,15 +535,17 @@ class BucketInfo
 
     package void clearResize ( size_t n )
     {
-        if (this.buckets.length != n)
-        {
-            this.buckets.length             = n;
-            this.bucket_list_indices.length = n;
+        this.buckets.length             = n;
+        this.bucket_list_indices.length = n;
 
-            if (this.n_filled > n)
-            {
-                this.n_filled = n;
-            }
+        /*
+         * this.n_filled must be adjusted for clear() to work because clear()
+         * resets all elements in this.filled_buckets, which is
+         * this.buckets[0 .. this.n_filled].
+         */
+        if (this.n_filled > n)
+        {
+            this.n_filled = n;
         }
 
         this.clear();
@@ -541,4 +576,64 @@ class BucketInfo
 
         Stderr('\n').flush();
     }
+}
+
+/*******************************************************************************
+
+    Verify bug #823 (empty buckets were reported to be not empty) is fixed
+
+*******************************************************************************/
+
+version (UnitTest):
+
+import ocean.core.Test;
+
+unittest
+{
+    auto info = new BucketInfo(3);
+
+    // Checks if the number of elements reported by info for each bucket is
+    // the expected value.
+
+    void checkNumElements ( int[] expected ... )
+    in
+    {
+        assert(expected.length == info.num_buckets);
+    }
+    body
+    {
+        foreach (i, n; expected)
+        {
+            // BucketInfo.opIndex(x) returns the number of elements in bucket x.
+            test!("==")(info[i], n);
+        }
+    }
+
+    checkNumElements(0, 0, 0);
+
+    // BucketInfo.put(x) increases the number of elements in bucket x by 1.
+
+    info.put(2);
+    checkNumElements(0, 0, 1);
+
+    info.put(0);
+    checkNumElements(1, 0, 1);
+
+    info.put(1);
+    checkNumElements(1, 1, 1);
+
+    info.clearResize(4);
+    checkNumElements(0, 0, 0, 0);
+
+    info.put(3);
+    checkNumElements(0, 0, 0, 1);
+
+    info.put(1);
+    checkNumElements(0, 1, 0, 1);
+
+    info.put(0);
+    checkNumElements(1, 1, 0, 1);
+
+    info.put(2);
+    checkNumElements(1, 1, 1, 1);
 }