+++ RELEASE_NOTES.md
@@ -51,6 +51,14 @@ Deprecations
   requires `toString` to return immutable strings which was not intended in
   this case.
 
+* `ocean.net.HttpRequest.getUint(T)(cstring, ref T, out bool)`
+
+  Deprecated in favor of `ocean.net.HttpRequest.getUnsigned(T)(cstring, ref T, out bool)`
+
+* `ocean.net.HttpRequest.getUint(T)(cstring, ref T)`
+
+  Deprecated in favor of `ocean.net.HttpRequest.getUnsigned(T)(cstring, ref T)`
+
 New Features
 ============